% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateOhdsiPalette.R
\name{createOhdsiPalette}
\alias{createOhdsiPalette}
\title{Create an OHDSI-themed color palette}
\usage{
createOhdsiPalette(
  seedColors = c("#336B91", "#69AED5", "#11A08A", "#FBC511", "#EB6622"),
  numColors = 5,
  adjustAmount = 0.1,
  darken = TRUE
)
}
\arguments{
\item{seedColors}{A character vector of colors to use as the seed for the palette. Defaults to a pre-defined set of colors.}

\item{numColors}{The number of colors to generate in the palette. The default is set to the number of unique age groups in a dataset.}

\item{adjustAmount}{The amount by which to adjust the brightness or saturation of the colors.}

\item{darken}{A logical value indicating whether to darken (TRUE) or lighten (FALSE) the colors.}
}
\value{
A character vector representing the color palette.
}
\description{
This function generates a color palette based on the provided seed colors, which are then interpolated to create a specified number of colors.
The palette can either be darkened or lightened based on user preference.
}
\examples{
createOhdsiPalette()  # Default usage with pre-defined colors and darkening effect
createOhdsiPalette(seedColors = c("#FF0000", "#00FF00", "#0000FF"), numColors = 5, darken = FALSE)
}
