% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugExposure-package.R, R/DrugExposure.R
\docType{package}
\name{OhdsiRPlots-package}
\alias{OhdsiRPlots-package}
\title{OhdsiRPlots: OhdsiRPlots: Some basic plots customized to OHDSI themes.}
\description{
The OhdsiRPlots R package has some basic plots customized to OHDSI themes. Helps make plots for OHDSI studies be more uniform.

The OhdsiRPlots R package has some basic plots customized to OHDSI themes. Helps make plots for OHDSI studies be more uniform.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/OhdsiRPlots/}
  \item \url{https://github.com/OHDSI/OhdsiRPlots}
  \item Report bugs at \url{https://github.com/OHDSI/OhdsiRPlots/issues}
}


Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/OhdsiRPlots/}
  \item \url{https://github.com/OHDSI/OhdsiRPlots}
  \item Report bugs at \url{https://github.com/OHDSI/OhdsiRPlots/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
